Create Spectra-Data.m
The observed Hydrogen-alpha line of HD 94028 has a longer wavelength than the laboratory value of 656.28 nm. This suggests that there may be external factors causing the deviation from the expected wavelength, such as interstellar gas and dust, Doppler effect, or instrumental errors. More research is necessary to determine the cause of the observed deviation. The given data shows a range of wavelengths from 630.02 nm to 679.86 nm.